work around raspbian not having completion.zsh.
it is included somewhere else, see https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=928392#15

Dear Maintainer,

I notice that this package copies key-bindings.zsh and completion.zsh
to the following locations:
- /usr/share/doc/fzf/examples/key-bindings.zsh
- /usr/share/zsh/vendor-completions/_fzf

However, in Arch, the files are copied to the following, different,
locations:
- /usr/share/fzf/key-bindings.zsh
- /usr/share/fzf/completion.zsh

I don't believe that completion.zsh is a valid zsh completion file, but
is instead intended to be sourced.

Would it be possible to update the package to match the Arch
installation locations, and update the README to instruct the user to
source both files?